requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.9
decorator==5.1.1
dill==0.3.8
dm-tree==0.1.8
docstring_parser==0.16
etils==1.11.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.9.0
googleapis-common-protos==1.66.0
huggingface-hub==0.26.3
idna==3.10
imageio==2.36.1
immutabledict==4.2.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.3
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest_asyncio==1.6.0
networkx==3.2.1
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
parso==0.8.4
pickleshare==0.7.5
pillow==11.0.0
pip==24.2
platformdirs==4.3.6
promise==2.3
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.29.1
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.1.0
pycocotools==2.0.8
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==307
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.1.0
simple-parsing==0.1.6
six==1.16.0
stack-data==0.6.2
sympy==1.13.1
tensorflow-metadata==1.16.1
termcolor==2.5.0
tfds-nightly==4.9.7.dev202412110044
tifffile==2024.9.20
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
transforms==0.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
wheel==0.44.0
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.1
zipp==3.21.0
